Account for corner case
Updated to account for removing the last entry from the cACertificate attribute of the NTAuth object. Previously, if the script attempted to delete the last value from the attribute, the DC would respond with an ObjectClassViolation error as the attribute is mandatory. Instead, it now detects when it attempts to remove the final value and instead replaces it with a null-byte.